+++ src/js/class-name.js
@@ -1,0 +1,4 @@
+// Exports:
+export function className(type) {
+  return type.prototype.modelName || type.prototype.constructor.name
+}