+++ src/js/get-by-id.js
@@ -1,0 +1,8 @@
+// Imports:
+import * as cache from './cache'
+import { className } from './class-name'
+
+// Exports:
+export function getById(type, query) {  
+  return cache.get(className(type), query)
+}