+++ src/js/instance-of.js
@@ -1,0 +1,4 @@
+// Exports:
+export function instanceOf(instance) {  
+  return instance.constructor.prototype.modelName || instance.constructor.name
+}