+++ test/js/model.dist.test.js
@@ -1,0 +1,6 @@
+// Imports:
+import Model from '../../dist/js/model'
+import { tests } from './tests'
+
+// Context:
+tests(Model)